fix: disable pixi-lock pre-commit hook in CI
If we want to run this in CI with the pre-commit.ci GitHub app, it seems
like we would need to make a separate repo where we can host the hooks.

I tried using various options for `languages` and
`additional_dependencies` to get a pixi binary installed locally but it
seems like it's a bit involved.